:sparkles: feat(backend): member resource에 있는 member cover-image 업로드 medias controller로 이동 (#273)
